tweak parameter;PyAudio context manager

+++ project1/main.py
@@ -1,18 +1,18 @@
 import wave
+from contextlib import contextmanager
 
 from pyaudio import PyAudio, paInt16
 
 CHUNK = 1024
 FORMAT = paInt16
 CHANNELS = 1
-RATE = 44100
-RECORD_SECONDS = 5
+RATE = 16000
+RECORD_SECONDS = 1
 
 
 def main():
     """Record 5sec into `output.wav`"""
-    with wave.open("output.wav", "wb") as file:
-        py_audio = PyAudio()
+    with wave.open("output.wav", "wb") as file, open_pyaudio() as py_audio:
         file.setnchannels(CHANNELS)
         file.setsampwidth(py_audio.get_sample_size(FORMAT))
         file.setframerate(RATE)
@@ -25,7 +25,15 @@ def main():
         print("Done")
 
         stream.close()
-        py_audio.terminate()
+
+
+@contextmanager
+def open_pyaudio():
+    py_audio = None
+    try:
+        yield (py_audio := PyAudio())
+    finally:
+        py_audio.terminate() if py_audio else None
 
 
 main() if __name__ == "__main__" else None
